% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viz_utils.R
\name{add_edges_by_reversibility}
\alias{add_edges_by_reversibility}
\title{Add Edges By Reversibility}
\usage{
add_edges_by_reversibility(grob, edge_width)
}
\arguments{
\item{grob}{a ggplot2 grob}

\item{edge_width}{width of edges on graph}
}
\description{
Add Edges By Reversibility
}
