% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/netcontextr_utils.R
\name{annotate_genes}
\alias{annotate_genes}
\title{Annotate netcontextr graph with data from a list-col of tibbles}
\usage{
annotate_genes(netcontextr_graph, annotation_data, field_name)
}
\arguments{
\item{netcontextr_graph}{a netcontextr graph object with "genes" and
"interactions" elements, Gene ID needs to be in Ensembl format and specified
in a column called "ensembl_gene_id"}

\item{annotation_data}{annotation data containing "ensembl_gene_id" and
"field_name"}

\item{field_name}{the specific column in the annotation data to use for
annotating the genes in the netcontextr graph object. This column should be a
list of tibbles}
}
\value{
modified netcontextr graph object, with a list of tibbles named
"field_name" added on to the "genes" element (each tibble containing a single
column denoting all of the annotated values in "annotation_data$field_name"
(or a one row one column tibble with NA if none)). The modified netcontextr
graph can now be trimmed using the `trim_network_by_gene_attribute` function
}
\description{
`annotate_genes` modifies a netcontextr graph by adding annotation data to
the "genes" object. Currently this function assumes that the "field_name" in
"annotation_data" is a list of tibbles. This function calls the helper
function `extract_gene_annotations` which assumes that "annotation_data" has
a column called "ensembl_gene_id".
}
\examples{
# annotate the genes in a netcontextr graph with subcellular localization(s)
# add this information to a new field called "compartment"
annotated_string_graph <- annotate_genes(
  example_string_graph,
  example_protein_subcellular_localizations,
  "compartment"
  )

# annotate the genes with expression data from GTEX
# first determine gene expression by tissue using zFPKM
gtex_expression_data_zFPKM <- gene_expression_by_tissue(example_gtex_data)
# second annotate genes with tissue expression(s)
# add this information to a new field called "tissue"
annotated_string_graph <- annotate_genes(example_string_graph,
  gtex_expression_data_zFPKM, "tissue")
}
