% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_basic_info.R
\name{basicInfoInput}
\alias{basicInfoInput}
\title{Basic Info Input}
\usage{
basicInfoInput(id, gui_label)
}
\arguments{
\item{id}{An ID string that corresponds with the ID used to call the module's
UI function.}

\item{gui_label}{GUI choice label}
}
\value{
Nothing; used for side-effects
}
\description{
UI components for the basic info shiny module
}
