% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_basic_info.R
\name{basicInfoServer}
\alias{basicInfoServer}
\title{Basic Info Server}
\usage{
basicInfoServer(id, species_names, species_identifiers, sbml_dfs, napistu)
}
\arguments{
\item{id}{An ID string that corresponds with the ID used to call the module's
UI function.}

\item{species_names}{A tibble containing names of all reaction species.}

\item{species_identifiers}{A tibble containing systematic identifiers for all reaction species nested by ontology.}

\item{sbml_dfs}{A python pathway object with class napistu.sbml_dfs_core.SBML_dfs}

\item{napistu}{reticulate connection to the napistu python package}
}
\value{
Nothing; used for side-effects
}
\description{
Server-side components for basic info shiny module
}
