% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bioconductor_organisms.R
\name{bioconductor_org_function}
\alias{bioconductor_org_function}
\title{Bioconductor Org Function}
\usage{
bioconductor_org_function(object_type, species)
}
\arguments{
\item{object_type}{Type of function to call}

\item{species}{Species name}
}
\value{
a tibble for function calls which can be coerced to a tabular result or the raw results otherwise
}
\description{
Call a function from one of the bioconductor org packages
}
\examples{
\dontrun{
bioconductor_org_function("ENZYME", "Homo sapiens")
}
}
