% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bioconductor_organisms.R
\name{bioconductor_org_library}
\alias{bioconductor_org_library}
\title{Bioconductor Org Library}
\usage{
bioconductor_org_library(species)
}
\arguments{
\item{species}{Species name}
}
\value{
0 invisibly
}
\description{
Load the organism bioconductor package for the specified species
}
\examples{

if (interactive()) {
  bioconductor_org_library("Homo sapiens")
}
}
