% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bioconductor_organisms.R
\name{bioconductor_org_package_prefix}
\alias{bioconductor_org_package_prefix}
\title{Bioconductor Org Package Prefix}
\usage{
bioconductor_org_package_prefix(species)
}
\arguments{
\item{species}{Species name}
}
\value{
character prefix
}
\description{
Find the prefix of the bioconductor package and its functions for a species
}
\examples{
bioconductor_org_package_prefix("Homo sapiens")
}
