% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config_default_python.R
\name{cleanup_napistu}
\alias{cleanup_napistu}
\title{Clean Up Napistu Environment}
\usage{
cleanup_napistu(napistu_list, force = FALSE)
}
\arguments{
\item{napistu_list}{Napistu environment object}

\item{force}{Logical, whether to force removal without confirmation}
}
\description{
Removes conda environment and miniconda if they were created by napistu
}
