% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config_default_python.R
\name{cleanup_napistu_env}
\alias{cleanup_napistu_env}
\title{Clean Up Napistu Environment}
\usage{
cleanup_napistu_env(napistu_env, force = FALSE)
}
\arguments{
\item{force}{Logical, whether to force removal without confirmation}

\item{env}{Napistu environment object}
}
\description{
Removes conda environment and miniconda if they were created by napistu
}
