% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/netcontextr_context.R
\name{compartmentalize_string}
\alias{compartmentalize_string}
\title{Compartmentalize STRING}
\usage{
compartmentalize_string(stringNetwork, aliases, hpa_localization_data)
}
\arguments{
\item{stringNetwork}{a flat text file of interactions downloaded from string}

\item{aliases}{key for converting between string protein ID and primary ID}

\item{hpa_localization_data}{path to HPA subcellular localization data}
}
\value{
a named list:
\describe{
  \item{compartmentalized_genes}{a tibble containing ensembl_gene and the comparments where the gene is expressed and interacts}
  \item{compartmentalized_edges}{an edgelist of interactions between pairs of genes and the compartments where these interactions could occur}
  \item{compartments}{a tibble of compartment meta-data}
  }
}
\description{
Apply netcontextr to compartmentalize the STRING interaction network
based on sub-cellular localization data from the human proteome atlas
}
\examples{
if (interactive()) {

  stringNetwork = "/group/cpr/external_pathways/string/9606.protein.links.full.v11.5.txt"
  aliases = "/group/cpr/external_pathways/string/9606.protein.aliases.v11.5.txt.gz"
  hpa_localization_data = "/group/cpr/external_pathways/protein_atlas/subcellular_location.tsv"

  compartmentalize_string(
    stringNetwork,
    aliases,
    hpa_localization_data
    )
}
}
