% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config_default_python.R
\name{conda_env_exists}
\alias{conda_env_exists}
\title{Check if Conda Environment Exists}
\usage{
conda_env_exists(env_name, conda_exe)
}
\arguments{
\item{env_name}{Name of conda environment}

\item{conda_exe}{Path to conda executable}
}
\value{
Logical indicating if environment exists
}
\description{
Check if Conda Environment Exists
}
