% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/python_utils.R
\name{convert_napistu_graph}
\alias{convert_napistu_graph}
\title{Convert Napistu Graph}
\usage{
convert_napistu_graph(napistu_graph, napistu_list)
}
\arguments{
\item{napistu_graph}{a Python NapistuGraph object (a subclass of igraph.Graph)}

\item{napistu_list}{A list containing loaded assets and bindings to
  Python modules.

\describe{
    \item{sbml_dfs}{
        SBML_dfs - the core pathway representation of the Napistu Python library
        }
    \item{napistu_graph}{
        Network graph - a Python igraph subclass with Napistu-specific
        attributes and methods
        }
    \item{species_identifiers}{Species identifier mappings}
    \item{precomputed_distances}{optional, distances between species nodes}
    \item{species_names}{
         A tibble containing the names of all genes, proteins, molecules, etc
         }
    \item{identifiers_nest}{
         A tibble with one row per ontology and a nested tibble containing
         all the identifiers and their corresponding molecular species
         }
    \item{python_modules}{A named list of Python modules: `napistu`}
    \item{python_environment}{See \link{validate_python_environment}}
    \item{napistu_config}{
        A `napistu_config` object dictating how the `napistu_list` was initialized
    }
    \item{loaded_at}{A date-time object indicating when `napistu_list` was initialized}
  }}
}
\value{
an S3 r_napistu_graph object which is a list containing vertices
and edges.
}
\description{
Convert a Python `graph.Graph` object to a list of vertices and edges.
}
\examples{
setup_napistu_list(create_napistu_config())
convert_napistu_graph(napistu_list$napistu_graph, napistu_list)
}
