% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/python_utils.R
\name{convert_sbml_dfs}
\alias{convert_sbml_dfs}
\title{Convert SBML DFs}
\usage{
convert_sbml_dfs(sbml_dfs)
}
\arguments{
\item{sbml_dfs}{an sbml_dfs object, generally accessed from a `napistu_list`}
}
\value{
an S3 r_sbml_dfs object which is a list with one table for each of
SBML_dfs' core attributes (compartments, species, compartmentalized_species,
reactions, and reaction_species)
}
\description{
Convert a Python Napistu SBML_dfs pathway representation into R tables. This
will remove some attributes (identifiers and sources) and all the methods which
are essential for the sbml_dfs to function as an SBML_dfs obbject but
the core relational structure of the pathway can be more easily explored
after this conversion.
}
\examples{
setup_napistu_list(create_napistu_config())
convert_sbml_dfs(napistu_list$sbml_dfs)
}
