% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config_default_python.R
\name{create_conda_environment}
\alias{create_conda_environment}
\title{Create Conda Environment}
\usage{
create_conda_environment(env_name, conda_exe, python_version = "3.11")
}
\arguments{
\item{env_name}{Name of environment to create}

\item{conda_exe}{Path to conda executable}

\item{python_version}{Python version to install}
}
\value{
NULL (side effect: creates environment)
}
\description{
Create a conda environment with Napistu installed.
}
