% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config_default_python.R
\name{create_default_conda_env}
\alias{create_default_conda_env}
\title{Create Default Conda Environment}
\usage{
create_default_conda_env(
  env_name = NAPISTU_CONSTANTS$DEFAULT_CONDA_ENV_NAME,
  verbose = TRUE
)
}
\arguments{
\item{env_name}{The name of the to-be-created conda environment}

\item{verbose}{Logical scalar; if TRUE (default), prints informative messages
about the current operation progress. If FALSE, runs silently.}
}
\value{
List with environment info and metadata
}
\description{
Setup an conda environment with Napistu installed. This will install miniconda
if needed (with user permission) and create an environment if-needed in
an existing conda/miniconda installation or the on-the-fly installation. This
process can be undone using \link{cleanup_napistu}.
}
\keyword{internal}
