% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/netcontextr_context.R
\name{create_mock_reactions}
\alias{create_mock_reactions}
\title{Create mock reactions}
\usage{
create_mock_reactions(n_reactions = 1000)
}
\arguments{
\item{n_reactions}{Number of fake reactions to create}
}
\value{
a tibble of reactions and their reaction species
}
\description{
Create a \code{netcontextr_reactions} list for use in
\code{trim_reactions_by_gene_attribute} by generating FAKE reactions from
the genes in STRING.
}
\examples{
mock_reactions <- create_mock_reactions(10)

}
