% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{create_napistu_config}
\alias{create_napistu_config}
\title{Create Napistu Configuration}
\usage{
create_napistu_config(python_config = list(), assets_config = list())
}
\arguments{
\item{python_config}{List describing the Python environment to use as part of
  the `napistu_config`. This can be the system python, a virtual environment
  or conda. If `python_config` is an empty list, a miniconda environment and
  installation can be established on-the-fly.

  \describe{
       \item{virtual_env}{Optional, path to a virtual environment}
       \item{conda}{
           Optional, path to a conda executable. This can be found with
           `reticulate::conda_exe()`. If defined, the `conda_env_name` should
           also be defined.
       }
       \item{conda_env_name}{
           Optional, name of the conda environment to use. A suitable environment
           can be created using \link{create_conda_environment}.
       }
       \item{python}{
           Optional, path to a Python distribution.
       }
  }}

\item{assets_config}{List containing asset configuration settings as part of
the `napistu_config`. If this is an empty list, the `test_pathway` bundled
with `napistu-r` will be used.
  
\describe{
    \item{assets_dir}{
        relative to this directory.
    }
    \item{sbml_dfs}{
        Path to a pickled `SBML_dfs` object generated by the
        Napistu Python package. `SBML_df`s are self-contained relational
        representations of a pathway.
    }
    \item{napistu_graph}{
        Path to a pickled `NapistuGraph` object generated by the Napistu
        Python package. `NapistuGraph`s are a subclass of `igraph.Graph`.
        }
    \item{species_identifiers}{
        Path to a .tsv containing the systematic identifiers for all
        of the molecular species in the `SBML_dfs`.
    }
    \item{precomputed_distances}{
        Optional, path to a .parquet file containing pairwise distances
        between molecular species. Defining this can speed up the calculation
        of shortest paths and neighborhoods.
    }
}}
}
\value{
Configuration object
}
\description{
Create Napistu Configuration
}
\examples{
\dontrun{
# Direct configuration
napistu_config <- create_napistu_config(
  python_config = list(virtualenv = "/path/to/venv"),
  assets_config = list(
    sbml_dfs = "/path/to/sbml.pkl",
    napistu_graph = "/path/to/graph.pkl",
    species_identifiers = "/path/to/species.tsv"
  )
)

# Minimal configuration (uses defaults)
napistu_config <- create_napistu_config()
}
}
