% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neighborhoods.R
\name{create_neighborhood_summary_table}
\alias{create_neighborhood_summary_table}
\title{Create Neighborhood Summary Table}
\usage{
create_neighborhood_summary_table(neighborhood_table)
}
\arguments{
\item{neighborhood_table}{a tibble produced by \link{create_neighborhood_table}
containing one row per neighborhood with nested lists as attributes:
\describe{
    \item{sc_name}{A human readible name for the focal vertex}
    \item{s_id}{The internal unique species id of the focal vertex}
    \item{c_id}{The internal unique compartment id of the focal vertex}
    \item{sc_id}{The internal unique compartmentalized species id of the focal vertex}
    \item{sc_Source}{The Source object for the focal vertex}
    \item{vertices}{The vertices in the focal vertex's neighborhood}
    \item{edges}{The edges in the focal vertex's neighborhood}
    \item{reaction_sources}{The source pathways of the reaction vertices}
}}
}
\value{
an unnested table of vertices across all neighborhoods
}
\description{
Unnest the neighborhoods' vertices
}
\keyword{internal}
