% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neighborhoods.R
\name{create_neighborhood_table}
\alias{create_neighborhood_table}
\title{Create Neighborhood Table}
\usage{
create_neighborhood_table(
  species_id,
  sbml_dfs,
  napistu_graph,
  precomputed_distances = NULL,
  network_type = "downstream",
  max_steps = 3L,
  max_neighbors = 10L,
  napistu
)
}
\arguments{
\item{species_id}{species identifier for focal node}

\item{sbml_dfs}{network reconstruction}

\item{napistu_graph}{igraph model of network}

\item{precomputed_distances}{A table of precomputed distances between pairs of
molecular species. Or, NULL to calculate this on-the-fly (default).}

\item{network_type}{what type of neighborhood should be formed (ignored
if \code{napistu_graph} is undirected).
\describe{
  \item{downstream}{descendants of the focal node}
  \item{upstream}{ancestors of the focal node}
  \item{hourglass}{descendants and ancestors of focal node}
}}

\item{max_steps}{number of steps away from focal node allowed}

\item{max_neighbors}{prune to this number of upstream regulators and
downstream targets}

\item{napistu}{binding to napistu python library}
}
\description{
Create Neighborhood Table
}
\examples{

if (interactive()) {
  interactive_initialization_wrapper()
  species_id <- "S00000061"

  create_neighborhood_table(
    species_id,
    sbml_dfs,
    napistu_graph,
    network_type = "hourglass",
    napistu = napistu
  )

 # calculating distances on-the-fly
  create_neighborhood_table(
    species_id,
    sbml_dfs,
    napistu_graph,
    precomputed_distances = precomputed_distances,
    network_type = "hourglass",
    napistu = napistu
  )
}
}
