% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/netcontextr_utils.R
\name{createStringGraph}
\alias{createStringGraph}
\title{Create igraph object from String network data}
\usage{
createStringGraph(
  stringNetwork = "data-raw/9606.protein.links.full.v11.5.txt.gz",
  primaryID = "Ensembl_gene",
  aliases = "data-raw/9606.protein.aliases.v11.5.txt.gz"
)
}
\arguments{
\item{stringNetwork}{a flat text file of interactions downloaded from string}

\item{primaryID}{desired primary ID to use in created igraph object,
by default this is Ensembl gene ID}

\item{aliases}{key for converting between string protein ID and primary ID}
}
\value{
A netcontextr graph object, specifically, a list comprised of
"genes" (a tibble of the unique set of genes in the graph,
eventually this will also contain metadata for the genes), and
"interactions" (a multi-column tibble of the full set of interactions along
with the various weight metrics derived from string).
}
\description{
`createStringGraph` uses a flat text file of gene/protein interactions,
downloaded from string.db to create netcontextr graph object
}
\examples{
if(interactive()){
  stringNetwork = "data-raw/9606.protein.links.full.v11.5.txt.gz"
  primaryID = "Ensembl_gene"
  aliases = "data-raw/9606.protein.aliases.v11.5.txt.gz"

  createStringGraph(
    stringNetwork = stringNetwork,
    primaryID = primaryID,
    aliases = aliases
   )
}
}
