% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subgraph.R
\name{define_subgraphs}
\alias{define_subgraphs}
\title{Create Subgraph}
\usage{
define_subgraphs(napistu_list, subgraph_vertices, max_components = 4L)
}
\arguments{
\item{napistu_list}{A list containing loaded assets and bindings to
  Python modules.

\describe{
    \item{sbml_dfs}{
        SBML_dfs - the core pathway representation of the Napistu Python library
        }
    \item{napistu_graph}{
        Network graph - a Python igraph subclass with Napistu-specific
        attributes and methods
        }
    \item{species_identifiers}{Species identifier mappings}
    \item{precomputed_distances}{optional, distances between species nodes}
    \item{species_names}{
         A tibble containing the names of all genes, proteins, molecules, etc
         }
    \item{identifiers_nest}{
         A tibble with one row per ontology and a nested tibble containing
         all the identifiers and their corresponding molecular species
         }
    \item{python_modules}{A named list of Python modules: `napistu`}
    \item{python_environment}{See \link{validate_python_environment}}
    \item{napistu_config}{
        A `napistu_config` object dictating how the `napistu_list` was initialized
    }
    \item{loaded_at}{A date-time object indicating when `napistu_list` was initialized}
  }}

\item{subgraph_vertices}{the vertices to retain in the subgraph}

\item{max_components}{the number of components to return}
}
\value{
a list of the `max_components` largest weakly-connected components
}
\description{
Filter to an induced subgraph containing a set of vertices and their connections
}
\examples{
suppressPackageStartupMessages(library(dplyr))
setup_napistu_list(create_napistu_config())
subgraph_vertices <- sample(napistu_list$napistu_graph$vs["name"], 100)
define_subgraphs(napistu_list, subgraph_vertices, max_components = 2)
}
