% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/netcontextr_context.R
\name{detect_unexpressed}
\alias{detect_unexpressed}
\title{Filter Unexpressed}
\usage{
detect_unexpressed(
  expression_data,
  feature_defining_var,
  invalid_vars = NULL,
  zFPKM_cutoff = -3
)
}
\arguments{
\item{expression_data}{A tabular dataset with rows defining distinct features
and columns as samples or contexts.}

\item{feature_defining_var}{A variable in \code{expression_data} defining
a distinct feature}

\item{invalid_vars}{Variables in \code{expression_data} which should be dropped
before applying zFPKM because they are either non-numeric or invalid contexts.}

\item{zFPKM_cutoff}{zFPKM values about this cutoff will be considered as
meaningful expression.}
}
\value{
A list containing:
\describe{
  \item{expression_data_df}{\code{expression_data} formatted as a data.frame.}
  \item{zFPKM_scores}{\code{zFPKM_scores} \code{expression_data_df} with zFPKM applied.}
  \item{is_expressed}{Boolean data.frame of \code{zFPKM_scores} thresholded by \code{zFPKM_cutoff}.}
}
}
\description{
Filter a tabular table of feature (rows) by contexts (columns) using
  zFPKM::zFPKM to generate context-specificity scores.
}
