% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/netcontextr_context.R
\name{discretize_expression}
\alias{discretize_expression}
\title{Discretize Expression}
\usage{
discretize_expression(
  tall_expression_data,
  valid_abundance_vars,
  valid_feature_vars,
  valid_context_vars,
  zFPKM_cutoff = -3
)
}
\arguments{
\item{tall_expression_data}{Table with one row per feature x context pair.}

\item{valid_abundance_vars}{valid variables which can define the expression level; One and only one match is expected.}

\item{valid_feature_vars}{valid variables which can define a unique feature (e.g., gene); One and only one match is expected.}

\item{valid_context_vars}{valid variables which can define a unique expression context. 1+ matches are expected. If multiple matches are present then will be combined into a single context before applying \code{\link{tall_data_detect_unexpressed}}.}

\item{zFPKM_cutoff}{zFPKM values about this cutoff will be considered as
meaningful expression.}
}
\value{
tall_expression_data with two variables added:
\describe{
  \item{zFPKM_score}{
  z-score of context \code{abundance_var} relative to other contexts.
  },
  \item{is_expressed}{
  Thresholded zFPKM scores indicating whether a gene is expressed (1) or
  not expressed (0) in this context.
  }
}
}
\description{
Identify the relevant fields in \code{tall_expression_data} for applying \code{\link{tall_data_detect_unexpressed}} and then apply this function. This provides a flexible means of handling multiple datasets including those where expression context is jointly defined by multiple variables.
}
