% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{efo_diseases}
\alias{efo_diseases}
\title{EFO Diseases}
\format{
A tibble with 18,497 rows and 2 columns:
\describe{
  \item{id}{The EFO ID}
  \item{name}{The name of the EFO phenotype}
}
}
\usage{
efo_diseases
}
\description{
The Experimental Factor Ontology used for diseases and phenotypes by
  Open Targets.
}
\keyword{datasets}
