% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{ensure_absolute_path}
\alias{ensure_absolute_path}
\title{Convert a path to an absolute path}
\usage{
ensure_absolute_path(unresolved_path)
}
\arguments{
\item{unresolved_path}{Character string representing a file or directory path}
}
\value{
Character string with the absolute path
}
\description{
\itemize{
  \item If the path is already absolute (begins with "/"), it is returned as-is.
  \item If the path is relative to a home directory (begins with "~"), expand using
the home directory path
  \item If the path is relative, it is converted to an absolute path using normalizePath().
}
}
\keyword{internal}
