% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example_gtex_data}
\alias{example_gtex_data}
\title{Example RNA-Seq data (median TPM per tissue) from GTEX V8}
\format{
## `example_gtex_data`
A tibble with 11,471 rows and 11 columns:
\describe{
  \item{ensembl_gene_id}{Ensembl gene id, e.g. ENSG00000223972}
  \item{ensembl_geneTranscript_id}{GTEX's hybrid gene/transcript ID}
  \item{Decription}{human gene symbol, e.g. DDX11L1}
  \item{Liver}{median TPM for Liver}
  \item{Spleen}{median TPM for Spleen}
  \item{Lung}{median TPM for Lung}
  \item{Testis}{median TPM for Testis}
  \item{Ovary}{median TPM for Ovary}
  \item{Whole.Blood}{median TPM for Whole.Blood}
  \item{Kidney...Cortex}{median TPM for Kidney...Cortex}
  \item{Kidney...Medulla}{median TPM for Kidney...Medulla}
}
}
\source{
<https://www.gtexportal.org/home/datasets>
}
\usage{
example_gtex_data
}
\description{
These are Analysis V8 data (dbGaP Accession phs000424.v8.p2) and correspond to
gene level median TPMs per tissue (54 tissues) summarized from ~17,000 samples.
}
\details{
NB these data were trimmed to genes present in example_string_graph as well
as to 8 tissue to reduce example dataset size.
}
\keyword{datasets}
