% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example_protein_subcellular_localizations}
\alias{example_protein_subcellular_localizations}
\title{Subcellular localization data from The Protein Atlas}
\format{
## `example_protein_subcellular_localizations`
A tibble with 13,105 rows and 15 columns:
\describe{
  \item{Gene}{Ensembl gene identifier}
  \item{Gene name}{name of gene}
  \item{Reliability}{gene reliability score}
  \item{Main Location}{Primary location(s) of the protein}
  \item{Additional Location}{Secondary location(s) of the protein}
  \item{Extracellular Location}{i.e. is the protein secreted}
  \item{Enhanced}{enhanced locations (strongest evidence)}
  \item{Supported}{supported locations (second strongest evidence)}
  \item{Approved}{approved locations (third strongest evidence)}
  \item{Uncertain}{uncertain locations (weakest evidence)}
  \item{Single-cell variation intensity}{locations with single-cell variation in intensity}
  \item{Single-cell variation spatial}{locations with spatial single-cell variation}
  \item{Cell cycle dependency}{locations with observed cell cycle dependency (type can be one or more of biological definition, custom data or correlation)}
  \item{GO id}{Gene Ontology Cellular Component term identifier}
  \item{compartment}{list-col containing a one column tibble with all of the subcompartments a given gene is observed in}
}
}
\source{
<https://www.proteinatlas.org/about/downloads>
}
\usage{
example_protein_subcellular_localizations
}
\description{
Subcellular location of proteins based on immunofluorescently stained cells.
The data is based on The Human Protein Atlas version 21.1 and Ensembl version 103.38.
See https://www.proteinatlas.org/about/assays+annotation#ihk for more
information on how these data were annotated / derived. Some specifics are
included below for reference.
}
\details{
The Subcellular section of the Human Protein Atlas provides high-resolution
insights into the expression and spatiotemporal distribution of proteins
encoded by 13041 genes (65% of the human protein-coding genes). For each gene,
the subcellular distribution of the protein has been investigated by
immunofluorescence (ICC-IF) and confocal microscopy in up to three different
cell lines, selected from a subset of 36 of the cell lines found in the Cell
Line Section. Upon image analysis, the subcellular localization of the protein
has been classified into one or more of 35 different organelles and fine
subcellular structures. In addition, the section includes an annotation of
genes that display single-cell variation in protein expression levels and/or
subcellular distribution, as well as an extended analysis of cell cycle
dependency of such variations. See also key publication:
Thul PJ et al. (2017) “A subcellular map of the human proteome” Science 356
https://pubmed.ncbi.nlm.nih.gov/28495876/

Annotation
In order to provide an interpretation of the staining patterns, all images
generated by ICC-IF are manually annotated. For each cell line and antibody,
the staining is described in terms of intensity, subcellular location and
single-cell variability (SCV). The staining intensity is classified as
negative, weak, moderate or strong based on the detector gain settings used
for image acquisition in combination with the visual appearance of the image.
SCVs within an immunofluorescence image are classified as intensity variation
(variation in their expression level) or as spatial variation (variation in
the spatial distribution).

Knowledge-based Annotation
The knowledge-based annotation aims to provide an interpretation of the
detected subcellular localization of a protein. In the first step, stainings
in different cell lines with the same antibody are reviewed and the results
are compared with external experimental protein/gene characterization data
for subcellular localization, available in the UniProtKB/Swiss-Prot database.
In the second step, all antibodies targeting the same protein are taken in
consideration for a final annotation of the subcellular distribution of the
protein.

Reliability Score
Each location is separately given one of the four reliability scores
(Enhanced, Supported, Approved, or Uncertain) based on available
protein/RNA/gene characterization data from both HPA and the
UniProtKB/Swiss-Prot database. The reliability score also encompasses several
additional factors, including reproducibility of the antibody staining in
different cell lines, correlation between staining intensity and RNA
expression levels, and assays for enhanced antibody validation. Enhanced
validation in achieved by using antibodies binding to different epitopes on
the same target protein (independent antibody validation), by assessing
staining intensity upon knockdown/knockout of the target protein
(genetic validation) and/or by matching of the signal with a GFP-tagged
protein (recombinant expression validation), and experimental evidence for
subcellular location described in literature. The individual location
relibility scores are summarized in an overall gene reliability score.

There are four different reliability scores:
Enhanced - The antibody has enhanced validation and there is no contradicting
data, such as literature describing experimental evidence for a different
location.
Supported - There is no enhanced validation of the antibody, but the
annotated localization is reported in literature.
Approved - The localization of the protein has not been previously described
and was detected by only one antibody without additional antibody validation.
Uncertain - The antibody-staining pattern contradicts experimental data or
expression is not detected at RNA level.

From Thul PJ et al. (2017) “A subcellular map of the human proteome” Science
main and additional locations are generated for each protein on the basis of
a clear difference either in the signal strength or in the occurrence across
the tested cell lines.

NB these data were not trimmed at all
}
\keyword{datasets}
