% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example_string_graph}
\alias{example_string_graph}
\title{Example protein network data from string.db}
\format{
## `example_string_graph`
A netcontextr graph object with 128 genes and 13,050 interactions.
Specifically, a list comprised of "genes" (a tibble of the unique set of
genes in the graph, eventually this will also contain metadata for the genes),
and "interactions" (a multi-column tibble of the full set of interactions
along with the various weight metrics derived from string (ranging from 0-1)).
\describe{
  \item{genes}{one column tibble of all genes in graph in ensembl gene id format}
  \item{interactions}{16 column tibble of all interactions in graph, first two
  columns are "gene 1" and "gene 2" in Ensembl gene id format, followed by
  14 different measures of interaction "score"}
}
}
\source{
<https://string-db.org/cgi/download.pl>
}
\usage{
example_string_graph
}
\description{
A subset of "full" protein links between human proteins, from string.db
NB full network means data includes distinction of direct vs. interologs
NB interologs = interolog is a conserved interaction between a pair of
proteins which have interacting homologs in another organism.
NB For most types of evidence, there are two types of scores:
the 'normal' score, and the 'transferred' score. The latter is computed from
data that is not originally observed in the organism of interest, but instead
in some other organism and then transferred via homology/orthology. All
potential source organisms are searched for evidence, but the actual transfers
to the receiving organism are made non-redundant (according to 'clades' of
closely related organisms in the tree of life).
}
\details{
From String paper at https://academic.oup.com/nar/article/49/D1/D605/6006194

All interaction evidence that contributes to a given network is benchmarked
and scored, and the scores are integrated into a final ‘combined score’. This
score is scaled between zero and one and provides an estimate of STRING’s
confidence on whether a proposed association is biologically meaningful given
all the contributing evidence. More specifically, The combined score is
computed by combining the probabilities from the different evidence channels
and corrected for the probability of randomly observing an interaction.
NB in the downloadable protein links files, the scores are multiplied by 1000
to make them integers.

Confidence limits are as follows:
low confidence - 0.15 (or better) | or 150 or better in protein links files
medium confidence - 0.4 | or 400 or better in protein links files
high confidence - 0.7 | or 700 or better in protein links files
highest confidence - 0.9 | or 900 or better in protein links files

Each association is provided as a connection
between two non-identical proteins, each from a different protein-coding gene
locus. STRING does not differentiate between splicing variants or
post-translationally modified protein isoforms encoded from the same locus—
instead, all such isoforms are collapsed and represented by a single,
canonical protein (i.e. a single protein per gene locus).

The various evidence types that contribute to STRING associations are first
benchmarked and scored separately, in seven distinct evidence channels.

The first three of the channels (neighborhood, fusion and co-occurrence)
contain computational association predictions based on whole-genome
comparisons. These so-called ‘genomic context’ channels can be computed for
all organisms for which a completely sequenced genome is available, and do
not depend on any further lab assays or measurements.
Neighborhood: two proteins are given an association score when their encoding
genes are in close proximity to each other on the chromosome. This channel is
applicable mostly for Bacteria and Archaea.
Fusion: STRING scans all genomes for open reading frames that appear to be
the result of gene-fusion events. For all inferred fusion events, the
constituent, non-fused genes in other genomes/organisms are given an
association score; the score is higher the better the fusion event can be
delineated in terms of the orthology of the participating genes. The
detection of gene fusion events across genomes can be used for predicting
functional associations of proteins, including physical interaction or
complex formation. See https://pubmed.ncbi.nlm.nih.gov/11820254/
Co-occurrence: STRING searches for pairs of genes whose occurrence patterns
throughout evolution show similarities. Such similarities can arise when
genes have been transferred, lost, or duplicated together during evolution,
which in turn can signify a shared function.

The next two channels are dealing with functional genomics experiments or
direct lab assays.
Co-expression: STRING is collecting gene expression evidence from a number
of sources; this is then normalized, pruned, and the expression profiles over
a large variety of conditions are compared. Pairs of genes that show
consistent similarities between their expression profiles are assigned
association scores; the majority of the expression data is RNA-based, but we
also import proteome expression data, from the ProteomeHD database.
Experiments: STRING collects protein–protein interaction evidence from
experiments and assays in the lab. This includes biochemical, biophysical and
genetic experiments; all such interaction evidence is imported from the
curated interaction database organized in the iMEX consortium, plus BioGRID.

The final two evidence channels deal with prior, consolidated knowledge on
protein–protein associations.
Database (AKA knowledge): STRING parses association evidence from curated
pathway databases, where it has been collected and consolidated manually by
expert curators. These include pathways annotated in KEGG, Reactome, and
MetaCyc, as well as protein complexes defined at the EBI Complex Portal or by
the Gene Ontology Consortium.
Textmining: STRING conducts statistical co-occurrence analysis across the
scientific literature. As of version 11.5 of STRING, the text-mining channel
is based on PubMed abstracts, articles from the PMC open access subset and
text from OMIM and SGD entry descriptions. Pairs of proteins mentioned
together in the same sentence, the same paragraph, or merely the same
publication are assigned a benchmarked association score, the calculation of
which is described in detail in (33).

NB the string interactions were trimmed to edges where both genes have
subcellular localization data (see example_protein_subcellular_localizations)
and edges involving a gene with genetic association data (see
example_genetic_association_data).
}
\keyword{datasets}
