% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/netcontextr_utils.R
\name{extract_gene_annotations}
\alias{extract_gene_annotations}
\title{Extract gene annotations from annotation data}
\usage{
extract_gene_annotations(gene, annotation_data, field_name)
}
\arguments{
\item{gene}{gene to be annotated, in ensembl ID format}

\item{annotation_data}{annotation data containing "ensembl_gene_id" and
"field_name"}

\item{field_name}{the specific column in the annotation data to use for
annotating the genes in the netcontextr graph object. This column should be a
list of tibbles}
}
\value{
a single column tibble containing all of the annotations for gene
in "annotation_data$field_name"
}
\description{
`extract_gene_annotations` matches a gene to the corresponding tibble in
"annotation_data$field_name". This is a helper function for `annotate_genes`,
and assumes that "annotation_data" has a column called "ensembl_gene_id".
}
\examples{
# add "compartment" annotation to a gene
example_gene_annotation <- extract_gene_annotations(example_string_graph$genes$gene[1],
  example_protein_subcellular_localizations, "compartment")
}
