% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config_assets.R
\name{get_bundled_asset_paths}
\alias{get_bundled_asset_paths}
\title{Get Bundled Asset Sources}
\usage{
get_bundled_asset_paths()
}
\value{
Named list of asset file paths for bundled assets
}
\description{
Get Bundled Asset Sources
}
\keyword{internal}
