% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config_default_python.R
\name{get_conda_installation}
\alias{get_conda_installation}
\title{Get Conda Installation}
\usage{
get_conda_installation()
}
\value{
Path to conda executable
}
\description{
Get Conda Installation
}
\keyword{internal}
