% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viz_utils.R
\name{get_layout_properties}
\alias{get_layout_properties}
\title{Get Layout Properties}
\usage{
get_layout_properties(network_layout = "fr")
}
\arguments{
\item{network_layout}{method to used for creating a network layout (e.g., `fr`, `kk`, `drl`)}
}
\value{
a list containing:
\describe{
    \item{layout_fxn}{The igraph function used for laying out the graph}
    \item{weights_attract}{TRUE if large weights pull vertices together; FALSE if large weights push them apart}
}
}
\description{
Get Layout Properties
}
\keyword{internal}
