% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interactive_env.R
\name{interactive_initialization}
\alias{interactive_initialization}
\title{Interactive Initialization}
\usage{
interactive_initialization(testing_env = FALSE)
}
\arguments{
\item{testing_env}{Work off a simple environment

Note that this approach breaks lots of conventions (assignment to global
environment; sourcing packages within functions, ...).}
}
\description{
Add objects which support MechNet to the global environment. This can be
  used to run the full app or test its interactive and non-interactive
  components.
}
