% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interactive_env.R
\name{interactive_initialization_wrapper}
\alias{interactive_initialization_wrapper}
\title{Interactive Initialization Wrapper}
\usage{
interactive_initialization_wrapper(testing_env = FALSE)
}
\arguments{
\item{testing_env}{Logical indicating whether to load assets for testing (TRUE)
or, prime-time pathway representations (FALSE)}
}
\value{
0, invisibly
}
\description{
Determine whether object created by interactive_initialization() are
  are present in the global environment, and if not call
  interactive_initialization()
}
