% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viz_utils.R
\name{label_vertices}
\alias{label_vertices}
\title{Label Vertices}
\usage{
label_vertices(
  vertices,
  max_labeled_species,
  node_types_to_label = "species",
  always_label = NULL
)
}
\arguments{
\item{vertices}{a table of the vertices to plot}

\item{max_labeled_species}{maximum number of species to label (to avoid
overplotting). Labels which are likely to overlap are removed based on the
graph's layout and the value of `target_plot_width`.}

\item{node_types_to_label}{what `node_type`s in `vertices` to consider for labeling}

\item{always_label}{always include these vertices regardless of their priority. Provide vertex names (generally starting with SC or R).}
}
\value{
vertices with a `ggtext_label` variable added
}
\description{
Takes a pre-sorted table of vertices and annotates the top entries with
visually appealing ggtext labels.
}
\keyword{internal}
