% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/netcontextr_utils.R
\name{load_and_clean_gtex_data}
\alias{load_and_clean_gtex_data}
\title{Load GTEX tissue specific expression data}
\usage{
load_and_clean_gtex_data(gtex_data)
}
\arguments{
\item{gtex_data}{path to GTEX tissue specific expression data (medians)}
}
\value{
a tibble containing all the information from the GTEX flat text file,
without modification (unlike related function `load_and_clean_hpa_data`)
}
\description{
`load_and_clean_gtex_data` loads tissue specific expression data from GTEX
(median value per gene per tissue).
}
\examples{
if(interactive()){
  # load gtex expression data
  gtex_file <- "data-raw/GTEx_Analysis_2017-06-05_v8_RNASeQCv1.1.9_gene_median_tpm.gct.gz"
  gtex_expression_data <- load_and_clean_gtex_data(gtex_file)
}
}
