% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/netcontextr_utils.R
\name{load_and_clean_hpa_data}
\alias{load_and_clean_hpa_data}
\title{Load and format Human Protein Atlas subcellular localization data}
\usage{
load_and_clean_hpa_data(hpa_localization_data)
}
\arguments{
\item{hpa_localization_data}{path to HPA subcellular localization data}
}
\value{
a tibble containing all of the information from the HPA flat text
 file plus an addition "compartment" column, which is a list of tibbles, with
 each tibble containing a single column denoting all of the annotated
 subcellular locations (or a one row one column tibble with NA if none)
}
\description{
`load_and_clean_hpa_data` loads subcellular localization data from the
Human Protein Atlas and then creates a "compartment" column which is a
list of tibbles (all of the localisations of each protein are denoted in the
single column tibble for that protein). The "compartment" column can be used
to annotate the genes in a netcontextr graph object using the
`annotate_genes` function. After annotation, the netcontextr graph can then
be trimmed using the `trim_network_by_gene_attribute` function
}
\examples{
if(interactive()){
  hpa_file <- "data-raw/subcellular_location.tsv.zip"
  hpa_localization_data <- load_and_clean_hpa_data(hpa_file)
}
}
