% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config_assets.R
\name{load_assets}
\alias{load_assets}
\title{Load Assets}
\usage{
load_assets(napistu_config, verbose = TRUE)
}
\arguments{
\item{napistu_config}{A napistu configuration object of class 'napistu_config'
containing all necessary configuration settings for the loading required
assets and configuring and/or using an exiting appropriate Python environment}

\item{verbose}{Logical scalar; if TRUE (default), prints informative messages
about the current operation progress. If FALSE, runs silently.}
}
\value{
A list containing loaded assets including:
  \describe{
    \item{sbml_dfs}{
        SBML_dfs - the core pathway representation of the Napistu Python library
        }
    \item{napistu_graph}{
        Network graph - a Python igraph subclass with Napistu-specific
        attributes and methods
        }
    \item{species_identifiers}{Species identifier mappings}
    \item{precomputed_distances}{optional, distances between species nodes}
    \item{species_names}{
         A tibble containing the names of all genes, proteins, molecules, etc
         }
    \item{identifiers_nest}{
         A tibble with one row per ontology and a nested tibble containing
         all the identifiers and their corresponding molecular species
         }
  }
}
\description{
Load a set of Napistu files
}
