% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{load_configured_assets}
\alias{load_configured_assets}
\title{Load Configured Assets}
\usage{
load_configured_assets(assets_config)
}
\arguments{
\item{assets_config}{Assets configuration list}
}
\value{
List of loaded assets
}
\description{
Load Configured Assets
}
