% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{load_napistu_config}
\alias{load_napistu_config}
\title{Load Napistu Configuration from YAML}
\usage{
load_napistu_config(file_path)
}
\arguments{
\item{path}{Path to YAML configuration file}
}
\value{
Configuration object
}
\description{
Load Napistu Configuration from YAML
}
