% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{load_single_asset}
\alias{load_single_asset}
\title{Load Single Asset File}
\usage{
load_single_asset(file_path, asset_type)
}
\arguments{
\item{file_path}{Path to asset file}

\item{asset_type}{Type of asset for appropriate loading method}
}
\value{
Loaded asset object
}
\description{
Load Single Asset File
}
