% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/napistu.R
\docType{package}
\name{napistu.r}
\alias{napistu.r-package}
\alias{napistu.r}
\title{\code{Napistu} R package}
\description{
R functions for the Napistu project supporting the creation, transformation, and network-based interrogation of biological pathways.
}
\author{
\strong{Maintainer}: Sean Hackett \email{seanmchackett@gmail.com} (\href{https://orcid.org/0000-0002-9553-4341}{ORCID})

Other contributors:
\itemize{
  \item Calico Life Sciences LLC [copyright holder, funder]
}

}
