% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_neighborhoods.R
\name{neighborhoodInput}
\alias{neighborhoodInput}
\title{Neighborhood Input}
\usage{
neighborhoodInput(id)
}
\arguments{
\item{id}{An ID string that corresponds with the ID used to call the module's
UI function.}
}
\value{
Nothing; used for side-effects
}
\description{
UI components for the neighborhood shiny module
}
