% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_neighborhoods.R
\name{neighborhoodServer}
\alias{neighborhoodServer}
\title{Neighborhood Server}
\usage{
neighborhoodServer(
  id,
  species_names,
  species_identifiers,
  sbml_dfs,
  napistu_graph,
  napistu
)
}
\arguments{
\item{id}{An ID string that corresponds with the ID used to call the module's
UI function.}

\item{species_names}{A tibble containing names of all reaction species.}

\item{species_identifiers}{A tibble containing systematic identifiers for all reaction species nested by ontology.}

\item{sbml_dfs}{A python pathway object with class napistu.sbml_dfs_core.SBML_dfs}

\item{napistu_graph}{A python igraph network of \code{sbml_dfs}}

\item{napistu}{reticulate connection to the napistu python package}
}
\value{
Nothing; used for side-effects
}
\description{
Server-side components of the neighborhood shiny module
}
