% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subgraph.R
\name{plot_one_component}
\alias{plot_one_component}
\title{Plot One Component}
\usage{
plot_one_component(
  napistu_list,
  component_list,
  score_overlay = NULL,
  score_label = NULL,
  score_palette = NULL,
  join_scores_on = "name",
  max_labeled_species = 20,
  network_layout = "fr",
  edge_width = 0.1,
  vertex_size = 6
)
}
\arguments{
\item{napistu_list}{A list containing loaded assets and bindings to
  Python modules.

\describe{
    \item{sbml_dfs}{
        SBML_dfs - the core pathway representation of the Napistu Python library
        }
    \item{napistu_graph}{
        Network graph - a Python igraph subclass with Napistu-specific
        attributes and methods
        }
    \item{species_identifiers}{Species identifier mappings}
    \item{precomputed_distances}{optional, distances between species nodes}
    \item{species_names}{
         A tibble containing the names of all genes, proteins, molecules, etc
         }
    \item{identifiers_nest}{
         A tibble with one row per ontology and a nested tibble containing
         all the identifiers and their corresponding molecular species
         }
    \item{python_modules}{A named list of Python modules: `napistu`}
    \item{python_environment}{See \link{validate_python_environment}}
    \item{napistu_config}{
        A `napistu_config` object dictating how the `napistu_list` was initialized
    }
    \item{loaded_at}{A date-time object indicating when `napistu_list` was initialized}
  }}

\item{component_list}{a list produced by \link{define_subgraphs} containing
the `component_graph` and `reaction_sources`}

\item{score_overlay}{optional, vertex-level scores containing `score`
and the merging attribute specified in `join_on`}

\item{score_label}{optional, name of disease being overlaid}

\item{score_palette}{optional, color palette for scores}

\item{join_scores_on}{variable to use when merging vertices and score}

\item{max_labeled_species}{the number of vertices to try to label. Some
labels may be dropped to improve clarity.}

\item{network_layout}{method to used for creating a network layout (e.g., `fr`, `kk`, `drl`)}

\item{edge_width}{width of edges on graph}

\item{vertex_size}{vertices' size}
}
\description{
Plot One Component
}
\examples{
suppressPackageStartupMessages(library(dplyr))
setup_napistu_list(create_napistu_config())
subgraph_vertices <- sample(napistu_list$napistu_graph$vs["name"], 100)
subgraph_list <- define_subgraphs(napistu_list, subgraph_vertices, max_components = 2)
component_list <- subgraph_list[[1]]
score_overlay <- tibble::tibble(name = sample(subgraph_vertices, 20)) \%>\%
    dplyr::mutate(score = abs(stats::rnorm(dplyr::n())))
plot_one_component(
    napistu_list,
    component_list,
    score_overlay = score_overlay,
    score_palette = "log2 fold-change"
)
}
