% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viz_utils.R
\name{prepare_rendering}
\alias{prepare_rendering}
\title{Prepare Rendering}
\usage{
prepare_rendering(
  r_graph,
  reaction_sources,
  network_layout,
  edge_weights,
  target_plot_width = 6
)
}
\arguments{
\item{r_graph}{an R igraph network}

\item{reaction_sources}{an optional mapping from `r_id` to `pathway_id` and `name`.}

\item{network_layout}{method to used for creating a network layout (e.g., `fr`, `kk`, `drl`)}

\item{target_plot_width}{optional, specification for how large the planned plot is.
knowing this helps to allow for more labels in a large plot where obscured
labels will be filtered.}
}
\value{
a list containing
\describe{
    \item{neighborhood_grob}{A ggplot `grob` with no aethetics or geoms added}
    \item{vertices_df}{A table of vertices with upstream metadata plus coordinates and updated labels}
    \item{pathway_coords}{The bounding box of reaction's assigned to each pathway}
}
}
\description{
Prepare for the network plot by laying out the pathway in a way that loosly
groups related vertices and by suppressed overplotted labels
}
\keyword{internal}
