% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viz_utils.R
\name{prepare_score_overlays}
\alias{prepare_score_overlays}
\title{Prepare Score Overlays}
\usage{
prepare_score_overlays(vertices, score_overlay = NULL, join_scores_on = "s_id")
}
\arguments{
\item{vertices}{a table of vertices containing the variable specified in `join_on`}

\item{score_overlay}{optional, vertex-level scores containing `score`
and the merging attribute specified in `join_on`}

\item{join_scores_on}{variable to use when merging vertices and score}
}
\description{
Merge a table of vertex-level data with a network representation.
}
\keyword{internal}
