% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/python_utils.R
\name{print.r_napistu_graph}
\alias{print.r_napistu_graph}
\title{Print method for r_napistu_graph objects}
\usage{
\method{print}{r_napistu_graph}(x, ...)
}
\arguments{
\item{x}{An r_napistu_graph object}

\item{...}{Additional arguments (not used)}
}
\description{
Print method for r_napistu_graph objects
}
