% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config_default_python.R
\name{process_required_modules_spec}
\alias{process_required_modules_spec}
\title{Process Required Modules Specification}
\usage{
process_required_modules_spec(required_modules)
}
\arguments{
\item{required_modules}{Named character vector of modules with optional version constraints}
}
\value{
Character vector of package specifications for conda/pip
}
\description{
Process Required Modules Specification
}
\keyword{internal}
