% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/open_targets_utils.R
\name{query_open_targets_indications}
\alias{query_open_targets_indications}
\title{Query Open Targets - Indications}
\usage{
query_open_targets_indications(target_ensembl_gene_ids, disease_id)
}
\arguments{
\item{target_ensembl_gene_ids}{a character vector of ensembl gene names}

\item{disease_id}{A disease to search from the EFO ontology}
}
\value{
a list of target attributes
}
\description{
Using Open Target's GraphQL API return indication score for a
  disease/phenotype of interest across a set of genes.
}
\examples{
target_ensembl_gene_ids <- c("ENSG00000090104", "ENSG00000163599")
disease_id <- "EFO_0001060" # celiac's disease
query_open_targets_indications(target_ensembl_gene_ids, disease_id)
}
