% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/open_targets.R
\name{query_open_targets_targets}
\alias{query_open_targets_targets}
\title{Query Open Targets - Targets}
\usage{
query_open_targets_targets(target_ensembl_gene_ids, post_n_genes = 10)
}
\arguments{
\item{target_ensembl_gene_ids}{a character vector of ensembl gene names}

\item{post_n_genes}{separate queries into posts of this number of genes}
}
\value{
a list of target attributes
}
\description{
Using Open Target's GraphQL API return gene-metadata and disease/phenotype
  associations for a set of genes.
}
\examples{
target_ensembl_gene_ids <- c("ENSG00000119718", "ENSG00000175354")
query_open_targets_targets(target_ensembl_gene_ids)
}
