% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{random_species}
\alias{random_species}
\title{Random Species}
\usage{
random_species(napistu_list)
}
\arguments{
\item{napistu_list}{A list containing loaded assets and bindings to
  Python modules.

\describe{
    \item{sbml_dfs}{
        SBML_dfs - the core pathway representation of the Napistu Python library
        }
    \item{napistu_graph}{
        Network graph - a Python igraph subclass with Napistu-specific
        attributes and methods
        }
    \item{species_identifiers}{Species identifier mappings}
    \item{precomputed_distances}{optional, distances between species nodes}
    \item{species_names}{
         A tibble containing the names of all genes, proteins, molecules, etc
         }
    \item{identifiers_nest}{
         A tibble with one row per ontology and a nested tibble containing
         all the identifiers and their corresponding molecular species
         }
    \item{python_modules}{A named list of Python modules: `napistu`}
    \item{python_environment}{See \link{validate_python_environment}}
    \item{napistu_config}{
        A `napistu_config` object dictating how the `napistu_list` was initialized
    }
    \item{loaded_at}{A date-time object indicating when `napistu_list` was initialized}
  }}
}
\value{
a species ID matching the primary key of the `species` table of the
SBML_dfs object.
}
\description{
Selects a random gene/molecule for demos
}
\examples{
setup_napistu_list(create_napistu_config(), overwrite = TRUE)
random_species(napistu_list)
}
\keyword{internal}
