% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcpr.R
\docType{package}
\name{rcpr}
\alias{rcpr-package}
\alias{rcpr}
\title{\code{rcpr} package}
\description{
R functions supporting Calico Pathway Resources (CPR)
}
\author{
\strong{Maintainer}: Sean Hackett \email{sean@calicolabs.com} (\href{https://orcid.org/0000-0002-9553-4341}{ORCID})

Other contributors:
\itemize{
  \item Calico Life Sciences LLC [copyright holder, funder]
}

}
