% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viz_utils.R
\name{select_score_overlay_palette}
\alias{select_score_overlay_palette}
\title{Select Score Overlay Palette}
\usage{
select_score_overlay_palette(score_palette, ...)
}
\arguments{
\item{score_palette}{the name of the palette to use
\describe{
    \item{indication_scores}{A palette going from [0,1]}
    \item{log2 fold-change}{A symmetric blue-black-yellow palette centered on zero}
}}

\item{...}{extra arguments to pass into the color palette}
}
\description{
Create a palette for coloring nodes based on vertex-level scores.
}
