% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_select_entity.R
\name{selectEntityInput}
\alias{selectEntityInput}
\title{Select Entity Input}
\usage{
selectEntityInput(id, gui_label)
}
\arguments{
\item{id}{An ID string that corresponds with the ID used to call the module's
UI function.}

\item{gui_label}{GUI choice label}
}
\value{
Nothing; used for side-effects
}
\description{
UI components for entity selection shiny module
}
