% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_select_entity.R
\name{selectEntityServer}
\alias{selectEntityServer}
\title{Select Entity Server}
\usage{
selectEntityServer(id, species_names, species_identifiers)
}
\arguments{
\item{id}{An ID string that corresponds with the ID used to call the module's
UI function.}

\item{species_names}{A tibble containing names of all reaction species.}

\item{species_identifiers}{A tibble containing systematic identifiers for all reaction species nested by ontology.}
}
\value{
Nothing; used for side-effects
}
\description{
Server-side components for the entity selection shiny module
}
