% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{setup_napistu}
\alias{setup_napistu}
\title{Complete Napistu Setup}
\usage{
setup_napistu(config)
}
\arguments{
\item{config}{Napistu configuration object or path to config file}
}
\value{
Napistu environment object with all components
}
\description{
Complete Napistu Setup
}
\examples{
\dontrun{
# Using configuration object
config <- napistu_config()
napistu_env <- setup_napistu(config)

# Using config file
napistu_env <- setup_napistu("my_config.yml")
}
}
