% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{setup_python_env}
\alias{setup_python_env}
\title{Setup Python Environment}
\usage{
setup_python_env(config)
}
\arguments{
\item{config}{Napistu configuration object}
}
\value{
List containing Python modules and environment info
}
\description{
Setup Python Environment
}
