% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_neighborhoods.R
\name{shiny_neighborhood_test}
\alias{shiny_neighborhood_test}
\title{Shiny Neighborhood Test}
\usage{
shiny_neighborhood_test(
  species_names,
  species_identifiers,
  sbml_dfs,
  napistu_graph,
  napistu
)
}
\arguments{
\item{species_names}{A tibble containing names of all reaction species.}

\item{species_identifiers}{A tibble containing systematic identifiers for all reaction species nested by ontology.}

\item{sbml_dfs}{A python pathway object with class napistu.sbml_dfs_core.SBML_dfs}

\item{napistu_graph}{A python igraph network of \code{sbml_dfs}}

\item{napistu}{reticulate connection to the napistu python package}
}
\value{
A shiny app
}
\description{
Demo for interrogating the neighborhood of an entity.
}
\examples{

if (interactive()) {
  interactive_initialization_wrapper()
  napistu <- reticulate::import("napistu")

  shiny_neighborhood_test(
    species_names,
    species_identifiers,
    sbml_dfs,
    napistu_graph,
    napistu = napistu
  )
}
}
