% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/open_targets_utils.R
\name{shiny_opentargets_test}
\alias{shiny_opentargets_test}
\title{Shiny Open Targets Test}
\usage{
shiny_opentargets_test()
}
\description{
A stand-alone app to test querying open targets using ensembl IDs
}
