% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_shortest_paths.R
\name{shortestPathsInput}
\alias{shortestPathsInput}
\title{Shortest Paths Input}
\usage{
shortestPathsInput(id)
}
\arguments{
\item{id}{An ID string that corresponds with the ID used to call the module's
UI function.}
}
\value{
Nothing; used for side-effects
}
\description{
UI components for the shortest paths shiny module
}
