% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viz_utils.R
\name{stub_grob}
\alias{stub_grob}
\title{Stub Grob}
\usage{
stub_grob(stub_str)
}
\arguments{
\item{stub_str}{the text to show in the stub}
}
\value{
the stubbed grob
}
\description{
Create a blank ggplot which can be interested in Shiny apps if a valid plot
can't be created
}
\keyword{internal}
