% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/open_targets.R
\name{summarize_open_targets_targets}
\alias{summarize_open_targets_targets}
\title{Summarize Open Targets - Targets}
\usage{
summarize_open_targets_targets(target_ensembl_gene_ids)
}
\arguments{
\item{target_ensembl_gene_ids}{a character vector of ensembl gene names}
}
\value{
a list containing:
\describe{
  \item{targets}{tibble summarizing each target and their known chemical
    modulators}
  \item{indications}{tibble summarizing disease/phenotypic associations
    scores related to each target. NULL if no scores are present}
}
}
\description{
Using Open Target's GraphQL API to read and format indications and
  tractability information for a set of genes.
}
\examples{
target_ensembl_gene_ids <- c(
  "ENSG00000999999",
  "ENSG00000119718",
  "ENSG00000175354",
  "ENSG00000091831"
  )
summarize_open_targets_targets(target_ensembl_gene_ids)
}
