% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/netcontextr_context.R
\name{tall_data_detect_unexpressed}
\alias{tall_data_detect_unexpressed}
\title{Tall Data Filter Unexpressed}
\usage{
tall_data_detect_unexpressed(
  tall_expression_data,
  feature_var,
  context_var,
  abundance_var,
  zFPKM_cutoff = -3
)
}
\arguments{
\item{tall_expression_data}{Table with one row per feature x context pair.}

\item{feature_var}{Variable describing a unique feature (e.g., a gene ID).}

\item{context_var}{Variable describing a unique context (e.g., a tissue or sample).}

\item{abundance_var}{Expression level which should be thresholded.}

\item{zFPKM_cutoff}{zFPKM values about this cutoff will be considered as
meaningful expression.}
}
\value{
tall_expression_data with two variables added:
\describe{
  \item{zFPKM_score}{
  Z-score of context \code{abundance_var} relative to other contexts.
  }
  \item{is_expressed}{
  Thresholded zFPKM scores indicating whether a gene is expressed (1) or
  not expressed (0) in this context.
  }
}
}
\description{
Format a "tall" table of features, contexts, and expression intensity and
  identify contexts where each feature is expressed based on zFPKM.
}
\examples{
if (interactive()) {
  gtex_file <- "data-raw/GTEx_Analysis_2017-06-05_v8_RNASeQCv1.1.9_gene_median_tpm.gct.gz"
  gtex_expression_data <- load_and_clean_gtex_data(gtex_file)

  gtex_tall_expression_data <- gtex_expression_data \%>\%
    dplyr::slice(1:200) \%>\%
    tidyr::gather(tissue, abundance, -c(ensembl_gene_id:Description))

  tall_data_detect_unexpressed(
    gtex_tall_expression_data, "ensembl_gene_id", "tissue", "abundance"
  )
}
}
