% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/netcontextr_context.R
\name{trim_network_by_edge_weight}
\alias{trim_network_by_edge_weight}
\title{Trim a netcontextr graph by edge weight}
\usage{
trim_network_by_edge_weight(netcontextr_graph, field_name, field_value)
}
\arguments{
\item{netcontextr_graph}{a netcontextr graph object with "genes" and
"interactions" elements}

\item{field_name}{the specific column in the "interactions" element of the
netcontextr graph to use for trimming interactions. This column should
contain some sort of numerical weight.}

\item{field_value}{the threshold for filtering (i.e. weight >= "field_value")}
}
\value{
modified netcontextr graph object, with "interactions" filtered to
 those with weight > "field_value" and genes filtered to those contained
 within the remaining interactions
}
\description{
`trim_network_by_edge_weight` trims a netcontextr graph object to weight >=
"field_value" as specified in the "field_name" column of the "interactions"
element in the netcontextr graph. For string interaction data, weights
range from 0-1 (see `createStringGraph`).
}
\details{
Confidence limits are as follows:
low confidence - 0.15 (or better) | or 150 or better in protein links files
medium confidence - 0.4 | or 400 or better in protein links files
high confidence - 0.7 | or 700 or better in protein links files
highest confidence - 0.9 | or 900 or better in protein links files
}
\examples{
# trim the edges of a netcontextr graph based on weight
high_confidence_string_graph <- trim_network_by_edge_weight(rcpr::example_string_graph,
  "combined_score", 0.7)
}
