% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/netcontextr_context.R
\name{trim_network_by_gene_attribute}
\alias{trim_network_by_gene_attribute}
\title{Trim a netcontextr graph based on gene metadata}
\usage{
trim_network_by_gene_attribute(
  netcontextr_graph,
  field_name,
  field_value = NA,
  addAttributeToInteractions = FALSE
)
}
\arguments{
\item{netcontextr_graph}{a netcontextr graph object with "genes" and
"interactions" elements}

\item{field_name}{the specific column in the "genes" element of the
netcontextr graph to use for trimming interactions. This column should be a
list of tibbles (i.e. as derived from `annotate_genes`)}

\item{field_value}{(optional) a specific value within "field_name" to filter
interactions on. "field_value" can be a vector of length one or greater.
If it is not specified (default) interactions are trimmed to those where
both nodes have overlapping values in "field_name". If "field_value" is
specified, then interactions are trimmed to those where both nodes have that
specific "field_value" within "field_name". The latter is useful for
trimming to a specific tissue (e.g. filter to interactions where both nodes
are expressed in a given tissue or tissues).}

\item{addAttributeToInteractions}{(optional) a logical (default = FALSE)
specifying whether to add field_name information as an interaction
attribute (memory intensive) or, remove the field_name column so it does
not become an interaction attribute. The default is to remove this column
because all of this information is already stored in the genes part of the
netcontextr graph object, and therefore, this information is redundant
(and results in the netcontextr graph object becoming quite large).}
}
\value{
modified netcontextr graph object, with "interactions" filtered to
 those where both genes/nodes share the same value(s) in "field_name". If
 "field_value" is specified, the "interactions" will be filtered to only
 those where both nodes have "field_value(s)" in "field_column". "Genes" are
 filtered to those contained within the remaining interactions. If
 addAttributeToInteractions is TRUE, field_name information will be added as
 an interaction attribute.
}
\description{
This function trims a netcontextr graph object to interactions where both
genes/nodes share the same value(s) in "field_name". Optionally, the user can
also specify a "field_value" to trim to interactions where both nodes have
that specific value in "field_name" only.
}
\examples{
# annotate the genes in a netcontextr graph with subcellular localization(s)
# add this information to a new field called "compartment"
annotated_string_graph <- annotate_genes(
  rcpr::example_string_graph,
  rcpr::example_protein_subcellular_localizations,
  "compartment"
  )
  
# now trim this network to interactions where nodes are found in overlapping compartments
compartmentalized_string_graph <- trim_network_by_gene_attribute(
  annotated_string_graph,
  "compartment"
  )

# Alternative example, where "field_value" is specified
# annotate the genes with expression data from GTEX
# first determine gene expression by tissue using zFPKM
gtex_expression_data_zFPKM <- gene_expression_by_tissue(rcpr::example_gtex_data)
# second annotate genes with tissue expression(s)
# add this information to a new field called "tissue"
annotated_string_graph <- annotate_genes(rcpr::example_string_graph,
  gtex_expression_data_zFPKM, "tissue")
# now trim network to interactions where both nodes are expressed in the liver
contextualized_string_graph <- trim_network_by_gene_attribute(
  annotated_string_graph,
  "tissue",
  "Liver"
  )
}
