% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/netcontextr_context.R
\name{trim_network_by_nodes}
\alias{trim_network_by_nodes}
\title{Trim Network by Nodes}
\usage{
trim_network_by_nodes(graph, nodesToKeep, requireBothNodes = TRUE)
}
\arguments{
\item{graph}{netcontextr graph object with "genes" and "interactions"}

\item{nodesToKeep}{character vector specifying the IDs of nodes to keep}

\item{requireBothNodes}{logical specifying whether to trim to interactions
where both nodesToKeep are present.}
}
\value{
netcontextr graph object, with nodes and edges trimmed to the
 specified nodes
}
\description{
Trim a netcontextr graph object by specifying the nodes to keep
}
\examples{
# trim example string graph to first 1000 proteins with subcellular localization data
# set requireBothNodes = TRUE (default) in order to restrict to interactions
# where both nodes are present within the first 1000 proteins
trimmed_graph <- trim_network_by_nodes(
  rcpr::example_string_graph,
  rcpr::example_protein_subcellular_localizations$ensembl_gene_id[1:1000],
  requireBothNodes = TRUE
  )

# trim example string graph to first 1000 proteins with subcellular localization data
# set requireBothNodes = FALSE in order to restrict to interactions where one
# or more nodes are present within the first 1000 proteins (this can be
# useful when elaborating the network of a key set of genes e.g. GWAS hits).
trimmed_graph <- trim_network_by_nodes(
  rcpr::example_string_graph,
  rcpr::example_protein_subcellular_localizations$ensembl_gene_id[1:1000],
  requireBothNodes = FALSE
  )
}
