% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/netcontextr_context.R
\name{trim_reactions_by_gene_attribute}
\alias{trim_reactions_by_gene_attribute}
\title{Trim Reactions By Gene Attribute}
\usage{
trim_reactions_by_gene_attribute(
  netcontextr_reactions,
  field_name,
  field_value = NA
)
}
\arguments{
\item{netcontextr_reactions}{a netcontextr reactions object with "genes" and
"reactions" elements}

\item{field_name}{the specific column in the "genes" element of the
netcontextr graph to use for trimming interactions. This column should be a
list of tibbles (i.e. as derived from `annotate_genes`)}

\item{field_value}{(optional) a specific value within "field_name" to filter
interactions on. "field_value" can be a vector of length one or greater.
If it is not specified (default) interactions are trimmed to those where
both nodes have overlapping values in "field_name". If "field_value" is
specified, then interactions are trimmed to those where both nodes have that
specific "field_value" within "field_name". The latter is useful for
trimming to a specific tissue (e.g. filter to interactions where both nodes
are expressed in a given tissue or tissues).}
}
\description{
Trim Reactions By Gene Attribute
}
\examples{
netcontextr_reactions <- create_mock_reactions(100)
annotated_reactions <- annotate_genes(
  netcontextr_reactions,
  rcpr::example_protein_subcellular_localizations,
  "compartment"
  )
 
trim_reactions_by_gene_attribute(annotated_reactions, "compartment")
}
