% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validators.R
\name{validate_asset_list}
\alias{validate_asset_list}
\title{Validate Assets List}
\usage{
validate_asset_list(asset_list)
}
\arguments{
\item{asset_list}{A list containing loaded assets including:
\describe{
    \item{sbml_dfs}{
        SBML_dfs - the core pathway representation of the Napistu Python library
        }
    \item{napistu_graph}{
        Network graph - a Python igraph subclass with Napistu-specific
        attributes and methods
        }
    \item{species_identifiers}{Species identifier mappings}
    \item{precomputed_distances}{optional, distances between species nodes}
    \item{species_names}{
         A tibble containing the names of all genes, proteins, molecules, etc
         }
    \item{identifiers_nest}{
         A tibble with one row per ontology and a nested tibble containing
         all the identifiers and their corresponding molecular species
         }
  }}
}
\value{
Invisible TRUE if valid, throws error if invalid
}
\description{
Validate Assets List
}
\keyword{internal}
