% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validators.R
\name{validate_asset_paths}
\alias{validate_asset_paths}
\title{Validate Asset Paths}
\usage{
validate_asset_paths(asset_paths)
}
\arguments{
\item{asset_paths}{List containing the paths of to-be-loaded assets}
}
\value{
Invisible TRUE if valid, throws error if invalid
}
\description{
Validate Asset Paths
}
\keyword{internal}
