% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validators.R
\name{validate_assets_config}
\alias{validate_assets_config}
\title{Validate Assets Config}
\usage{
validate_assets_config(assets_config, minimal_validation = FALSE)
}
\arguments{
\item{assets_config}{List containing asset configuration settings as part of
the `napistu_config`. If this is an empty list, the `test_pathway` bundled
with `napistu-r` will be used.
  
\describe{
    \item{assets_dir}{
        relative to this directory.
    }
    \item{sbml_dfs}{
        Path to a pickled `SBML_dfs` object generated by the
        Napistu Python package. `SBML_df`s are self-contained relational
        representations of a pathway.
    }
    \item{napistu_graph}{
        Path to a pickled `NapistuGraph` object generated by the Napistu
        Python package. `NapistuGraph`s are a subclass of `igraph.Graph`.
        }
    \item{species_identifiers}{
        Path to a .tsv containing the systematic identifiers for all
        of the molecular species in the `SBML_dfs`.
    }
    \item{precomputed_distances}{
        Optional, path to a .parquet file containing pairwise distances
        between molecular species. Defining this can speed up the calculation
        of shortest paths and neighborhoods.
    }
}}

\item{minimal_validation}{If TRUE then only the structure of `assets_config` will
be validated. If FALSE then the definition and existence of required assets
will be confirmed.}
}
\value{
Invisible TRUE if valid, throws error if invalid
}
\description{
Validate Assets Config
}
\keyword{internal}
