% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config_default_python.R
\name{validate_conda_executable}
\alias{validate_conda_executable}
\title{Validate Conda Executable}
\usage{
validate_conda_executable(conda_exe)
}
\arguments{
\item{conda_exe}{Path to conda executable}
}
\value{
Logical indicating if conda works
}
\description{
Validate Conda Executable
}
