% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validators.R
\name{validate_env_name}
\alias{validate_env_name}
\title{Validate Python Environment Name}
\usage{
validate_env_name(env_name)
}
\arguments{
\item{env_name}{The python environment name (a conda name or venv path)}
}
\value{
Invisible TURE if valid, throws error if invalid
}
\description{
Validate Python Environment Name
}
\keyword{internal}
