% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validators.R
\name{validate_env_type}
\alias{validate_env_type}
\title{Validate Python Environment Type}
\usage{
validate_env_type(validate_env_type)
}
\arguments{
\item{env_type}{Either "conda", "virtualenv", or "python"}
}
\value{
Invisible TURE if valid, throws error if invalid
}
\description{
Validate Python Environment Type
}
\keyword{internal}
