% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validators.R
\name{validate_max_labeled_species}
\alias{validate_max_labeled_species}
\title{Validate Max Labeled Species}
\usage{
validate_max_labeled_species(max_labeled_species)
}
\arguments{
\item{max_labeled_species}{maximum number of species to label (to avoid
overplotting). Labels which are likely to overlap are removed based on the
graph's layout and the value of `target_plot_width`.}
}
\value{
Invisible TRUE if valid, throws error if invalid
}
\description{
Validate Max Labeled Species
}
\keyword{internal}
