% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validators.R
\name{validate_neighborhood_table}
\alias{validate_neighborhood_table}
\title{Validate Neighborhood Table}
\usage{
validate_neighborhood_table(neighborhood_table)
}
\arguments{
\item{neighborhood_table}{a tibble produced by \link{create_neighborhood_table}
containing one row per neighborhood with nested lists as attributes:
\describe{
    \item{sc_name}{A human readible name for the focal vertex}
    \item{s_id}{The internal unique species id of the focal vertex}
    \item{c_id}{The internal unique compartment id of the focal vertex}
    \item{sc_id}{The internal unique compartmentalized species id of the focal vertex}
    \item{sc_Source}{The Source object for the focal vertex}
    \item{vertices}{The vertices in the focal vertex's neighborhood}
    \item{edges}{The edges in the focal vertex's neighborhood}
    \item{edge_source}{The source pathways of the reaction vertices}
}}
}
\value{
Invisible TRUE if valid, throws error if invalid
}
\description{
Determine whether the neighborhood table follows the expected structure.
Neighborhoods are a subnetwork of molecular species and reactions tightly connected to
a focal vertex.
}
\keyword{internal}
