% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validators.R
\name{validate_neighborhoods_network_type}
\alias{validate_neighborhoods_network_type}
\title{Validate Neighborhood Network Type}
\usage{
validate_neighborhoods_network_type(network_type)
}
\arguments{
\item{network_type}{what type of neighborhood should be formed (ignored
if \code{napistu_graph} is undirected).
\describe{
  \item{downstream}{descendants of the focal node}
  \item{upstream}{ancestors of the focal node}
  \item{hourglass}{descendants and ancestors of focal node}
}}
}
\value{
Invisible TRUE if valid, throws error if invalid
}
\description{
Validate Neighborhood Network Type
}
\keyword{internal}
