% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validators.R
\name{validate_network_layout}
\alias{validate_network_layout}
\title{Validate Network Layout}
\usage{
validate_network_layout(network_layout = "fr")
}
\arguments{
\item{network_layout}{method to used for creating a network layout (e.g., `fr`, `kk`, `drl`)}
}
\value{
Invisible TRUE if valid, throws error if invalid
}
\description{
Validate Network Layout
}
\keyword{internal}
