% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validators.R
\name{validate_overwrite}
\alias{validate_overwrite}
\title{Validate Verbose Parameter}
\usage{
validate_overwrite(overwrite)
}
\arguments{
\item{overwrite}{Overwrite existing caches}
}
\value{
Invisible TRUE if valid, throws error if invalid
}
\description{
Validate Verbose Parameter
}
\keyword{internal}
