% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validators.R
\name{validate_python_environment}
\alias{validate_python_environment}
\title{Validate Python Environment}
\usage{
validate_python_environment(python_environment)
}
\arguments{
\item{python_environment}{Summary of the loaded Python environment and specification for cleanup procedure used by \link{cleanup_napistu}
\describe{
    \item{path}{The environment being used}
    \item{type}{The type of reticulate configuration: "virtualenv", "conda", "python"}
    \item{created_by_napistu}{Was the environment created by Napistu? If so it can be removed by \link{cleanup_napistu}}
    \item{miniconda_installed}{Was Miniconda installed by Napistu? If so it can be removed by \link{cleanup_napistu}}
}}
}
\value{
Invisible TRUE if valid, throws error if invalid
}
\description{
Validate Python Environment
}
\keyword{internal}
