% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validators.R
\name{validate_python_list}
\alias{validate_python_list}
\title{Validate Python List}
\usage{
validate_python_list(python_list)
}
\arguments{
\item{python_list}{A list containing reticulate bindings to Python packages and summaries of the Python environment.
\describe{
    \item{python_modules}{A named list of Python modules: `napistu`}
    \item{python_environment}{See \link{validate_python_environment}}
}}
}
\value{
Invisible TRUE if valid, throws error if invalid
}
\description{
Validate Python List
}
\keyword{internal}
