% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validators.R
\name{validate_score_overlay_and_join_scores_on}
\alias{validate_score_overlay_and_join_scores_on}
\title{Validate Score Overlay and Join Scores On}
\usage{
validate_score_overlay_and_join_scores_on(score_overlay, join_scores_on)
}
\arguments{
\item{score_overlay}{optional, vertex-level scores containing `score`
and the merging attribute specified in `join_on`}

\item{join_scores_on}{variable to use when merging vertices and score}
}
\value{
Invisible TRUE if valid, throws error if invalid
}
\description{
Validate Score Overlay and Join Scores On
}
\keyword{internal}
