% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validators.R
\name{validate_score_palette}
\alias{validate_score_palette}
\title{Validate Score Palette}
\usage{
validate_score_palette(score_palette, score_overlay)
}
\arguments{
\item{score_palette}{optional, color palette for scores. If provided this can be a string defining built-in palettes or a custom palette
\describe{
    \item{log2 fold-change}{A blue -> black -> yellow color palette which is symmetric around zero}
    \item{indication scores}{A gray -> yellow -> orange -> red palette ranging from 0-1}
    \item{otherwise}{A `Scale` object defining a custom palette}
}}

\item{score_overlay}{optional, vertex-level scores containing `score`
and the merging attribute specified in `join_on`}
}
\value{
Invisible TRUE if valid, throws error if invalid
}
\description{
Validate Score Palette
}
\keyword{internal}
