% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validators.R
\name{validate_target_plot_width}
\alias{validate_target_plot_width}
\title{Validate Target Plot Width}
\usage{
validate_target_plot_width(target_plot_width)
}
\arguments{
\item{target_plot_width}{optional, specification for how large the planned plot is.
knowing this helps to allow for more labels in a large plot where obscured
labels will be filtered.}
}
\value{
Invisible TRUE if validation passes, otherwise throws an error
}
\description{
Validate Target Plot Width
}
\keyword{internal}
