% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validators.R
\name{validate_vertex_size}
\alias{validate_vertex_size}
\title{Validate Vertex Size}
\usage{
validate_vertex_size(vertex_size)
}
\arguments{
\item{vertex_size}{vertices' size}
}
\value{
Invisible TRUE if valid, throws error if invalid
}
\description{
Validate Vertex Size
}
\keyword{internal}
