% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_basic_info.R
\name{basicInfoServer}
\alias{basicInfoServer}
\title{Basic Info Server}
\usage{
basicInfoServer(id, species_names, species_identifiers, consensus_model, cpr)
}
\arguments{
\item{id}{An ID string that corresponds with the ID used to call the module's
UI function.}

\item{species_names}{A tibble containing names of all reaction species.}

\item{species_identifiers}{A tibble containing systematic identifiers for all reaction species nested by ontology.}

\item{consensus_model}{A python pathway object with class cpr.sbml.SBML_dfs}

\item{cpr}{reticulate connection to the cpr python package}
}
\value{
Nothing; used for side-effects
}
\description{
Server-side components for basic info shiny module
}
