% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config_python.R
\name{configure_existing_python}
\alias{configure_existing_python}
\title{Configure Existing Python Environment}
\usage{
configure_existing_python(python_config, verbose = TRUE)
}
\arguments{
\item{python_config}{Python configuration list}

\item{verbose}{if TRUE provide extra logging (default); FALSE to silence logging.}
}
\value{
List with environment path and metadata
}
\description{
Configure Existing Python Environment
}
