% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config_default_python.R
\name{create_default_conda_env}
\alias{create_default_conda_env}
\title{Create Default Conda Environment}
\usage{
create_default_conda_env(
  env_name = NAPISTU_CONSTANTS$DEFAULT_CONDA_ENV_NAME,
  verbose = TRUE
)
}
\arguments{
\item{env_name}{The name of the to-be-created conda environment}

\item{verbose}{if TRUE provide extra logging (default); FALSE to silence logging.}
}
\value{
List with environment info and metadata
}
\description{
Create Default Conda Environment
}
