% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{create_napistu_config}
\alias{create_napistu_config}
\title{Create Napistu Configuration}
\usage{
create_napistu_config(python_config = list(), assets_config = list())
}
\arguments{
\item{python_config}{List describing whether to use the system python,
virtual environment or conda and what environment to use if using an
environment manager. Typically extracted from the main napistu_config
object.}

\item{assets_config}{List containing asset configuration settings, typically
extracted from the main napistu_config object}
}
\value{
Configuration object
}
\description{
Create Napistu Configuration
}
\examples{
\dontrun{
# Direct configuration
napistu_config <- create_napistu_config(
  python_config = list(virtualenv = "/path/to/venv"),
  assets_config = list(
    sbml_dfs = "/path/to/sbml.pkl",
    napistu_graph = "/path/to/graph.pkl",
    species_identifiers = "/path/to/species.tsv"
  )
)

# Minimal configuration (uses defaults)
napistu_config <- create_napistu_config()
}
}
