% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neighborhoods.R
\name{create_neighborhood_table}
\alias{create_neighborhood_table}
\title{Create Neighborhood Table}
\usage{
create_neighborhood_table(
  napistu_list,
  species_id,
  network_type = "downstream",
  max_steps = 3L,
  max_neighbors = 10L
)
}
\arguments{
\item{napistu_list}{A list containing loaded assets and bindings to
  Python modules.

\describe{
    \item{sbml_dfs}{
        SBML_dfs - the core pathway representation of the Napistu Python library
        }
    \item{napistu_graph}{
        Network graph - a Python igraph subclass with Napistu-specific
        attributes and methods
        }
    \item{species_identifiers}{Species identifier mappings}
    \item{precomputed_distances}{optional, distances between species nodes}
    \item{species_names}{
         A tibble containing the names of all genes, proteins, molecules, etc
         }
    \item{identifiers_nest}{
         A tibble with one row per ontology and a nested tibble containing
         all the identifiers and their corresponding molecular species
         }
    \item{python_modules}{A named list of Python modules: `napistu`}
    \item{python_environment}{See \link{validate_python_environment}}
    \item{napistu_config}{
        A `napistu_config` object dictating how the `napistu_list` was initialized
    }
    \item{loaded_at}{A date-time object indicating when `napistu_list` was initialized}
  }}

\item{species_id}{species identifier for focal node}

\item{network_type}{what type of neighborhood should be formed (ignored
if \code{napistu_graph} is undirected).
\describe{
  \item{downstream}{descendants of the focal node}
  \item{upstream}{ancestors of the focal node}
  \item{hourglass}{descendants and ancestors of focal node}
}}

\item{max_steps}{number of steps away from focal node allowed}

\item{max_neighbors}{prune to this number of upstream regulators and
downstream targets}
}
\value{
a tibble containing one row per neighborhood with nested lists as attributes:
\describe{
    \item{sc_name}{A human readible name for the focal vertex}
    \item{s_id}{The internal unique species id of the focal vertex}
    \item{c_id}{The internal unique compartment id of the focal vertex}
    \item{sc_id}{The internal unique compartmentalized species id of the focal vertex}
    \item{sc_Source}{The Source object for the focal vertex}
    \item{vertices}{The vertices in the focal vertex's neighborhood}
    \item{edges}{The edges in the focal vertex's neighborhood}
    \item{edge_source}{The source pathways of the reaction vertices}
}
}
\description{
Create Neighborhood Table
}
\examples{
setup_napistu_list(create_napistu_config(), overwrite = TRUE)
species_id <- "S00000061"

create_neighborhood_table(
    species_id,
    napistu_list = napistu_list,
    network_type = "hourglass"
)
}
