% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{ensure_absolute_path}
\alias{ensure_absolute_path}
\title{Convert a path to an absolute path}
\usage{
ensure_absolute_path(unresolved_path)
}
\arguments{
\item{unresolved_path}{Character string representing a file or directory path}
}
\value{
Character string with the absolute path
}
\description{
If the path is already absolute (begins with "/"), it is returned as-is.
If the path is relative, it is converted to an absolute path using normalizePath().
}
\keyword{internal}
