% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config_assets.R
\name{load_assets_from_paths}
\alias{load_assets_from_paths}
\title{Load Assets from Paths}
\usage{
load_assets_from_paths(asset_paths, verbose = TRUE)
}
\arguments{
\item{asset_paths}{List containing the paths of to-be-loaded assets}

\item{verbose}{Logical scalar; if TRUE (default), prints informative messages
about the current operation progress. If FALSE, runs silently.}
}
\value{
List of loaded assets
}
\description{
Load Assets from Paths
}
