% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{load_napistu_config}
\alias{load_napistu_config}
\title{Load Napistu Configuration from YAML}
\usage{
load_napistu_config(file_path, verbose = TRUE)
}
\arguments{
\item{file_path}{Path to YAML configuration file}

\item{verbose}{Logical scalar; if TRUE (default), prints informative messages
about the current operation progress. If FALSE, runs silently.}
}
\value{
Configuration object
}
\description{
Load Napistu Configuration from YAML
}
