% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config_assets.R
\name{load_single_asset}
\alias{load_single_asset}
\title{Load Single Asset File}
\usage{
load_single_asset(file_path, asset_name)
}
\arguments{
\item{file_path}{Path to asset file}

\item{asset_name}{Name of asset for context}
}
\value{
Loaded asset object
}
\description{
Load Single Asset File
}
