% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config_assets.R
\name{load_single_asset}
\alias{load_single_asset}
\title{Load Single Asset File}
\usage{
load_single_asset(file_path, python_list, asset_name)
}
\arguments{
\item{file_path}{Path to asset file}

\item{python_list}{A list containing reticulate bindings to Python packages and summaries of the Python environment.
\describe{
    \item{python_modules}{A named list of Python modules: `napistu`}
    \item{python_environment}{See \link{validate_python_environment}}
}}

\item{asset_name}{Name of asset for context}
}
\value{
Loaded asset object
}
\description{
Load Single Asset File
}
