% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neighborhoods.R
\name{plot_one_neighborhood}
\alias{plot_one_neighborhood}
\title{Plot One Neighborhood}
\usage{
plot_one_neighborhood(
  vertices,
  edges,
  edge_sources,
  sc_id,
  sc_name,
  consensus_graph,
  score_overlay = NULL,
  score_label = NULL,
  score_palette = NULL,
  max_labeled_species = 30L,
  network_layout = "fr",
  edge_width = 0.1
)
}
\arguments{
\item{vertices}{table of species and reactions, produced by \link{create_neighborhood_table}}

\item{edges}{table of connections between species and reactions, produced by \link{create_neighborhood_table}}

\item{edge_sources}{table describing the model(s) each reaction comes from, produced by \link{create_neighborhood_table}}

\item{sc_id}{compartmentalized species identifier of focal node}

\item{sc_name}{name of focal node}

\item{consensus_graph}{igraph model of network}

\item{score_overlay}{optional, scores of neighbors for an indication from \link{summarize_indication}.}

\item{score_label}{optional, name of disease being overlaid.}

\item{score_palette}{optional, color palette for scores.}

\item{max_labeled_species}{maximum number of species to label (to avoid overplotting)}

\item{network_layout}{method to used for creating a network layout (e.g., fr, kk, drl)}

\item{edge_width}{width of edges on graph}
}
\value{
a ggplot2 grob
}
\description{
Plot One Neighborhood
}
\examples{

if (interactive()) {
  interactive_initialization_wrapper(TRUE)

  #species_id <- rownames(consensus_model$species)[
  #  consensus_model$species$s_name == "pyruvate kinase tetramer"
  #  ]
  species_id <- sample(rownames(consensus_model$species), 1)

  neighborhood_table <- create_neighborhood_table(
    species_id,
    consensus_model,
    consensus_graph,
    network_type = "hourglass",
    cpr = cpr,
    max_neighbors = 30L,
    max_steps = 15L
  )
  entry <- 1
  vertices <- neighborhood_table$vertices[[entry]]
  edges <- neighborhood_table$edges[[entry]]
  edge_sources <- neighborhood_table$edge_sources[[entry]]
  sc_id <- neighborhood_table$sc_id[entry]
  sc_name <- neighborhood_table$sc_name[entry]

  score_overlay <- vertices \%>\%
    dplyr::filter(node_type == "species") \%>\%
    dplyr::distinct(s_id) \%>\%
    dplyr::sample_frac(0.5) \%>\%
    dplyr::mutate(score = rnorm(n()))

  # score_overlay <- summarize_indication(
  #   disease_id = "EFO_0000400",
  #   consensus_model,
  #   create_neighborhood_summary_table(neighborhood_table),
  #   species_identifiers
  #   )

  plot_one_neighborhood(
    vertices,
    edges,
    edge_sources,
    sc_id,
    sc_name,
    consensus_graph,
    score_overlay = NULL
  )

  plot_one_neighborhood(
    vertices,
    edges,
    edge_sources,
    sc_id,
    sc_name,
    consensus_graph,
    score_overlay = score_overlay,
    score_palette = "log2 fold-change"
  )
}
}
