% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shortest_paths.R
\name{plot_shortest_path_network}
\alias{plot_shortest_path_network}
\title{Plot Shortest Path Network}
\usage{
plot_shortest_path_network(
  shortest_paths_list,
  consensus_graph,
  max_labeled_species = 10L
)
}
\arguments{
\item{shortest_paths_list}{results from napistu$network$find_all_shortest_reaction_paths}

\item{consensus_graph}{igraph model of network}

\item{max_labeled_species}{maximum number of species to label (to avoid overplotting)}
}
\description{
Plot Shortest Path Network
}
\examples{

if (interactive()) {
  interactive_initialization_wrapper()

  end_points <- sample(rownames(species_names), 2)
  source_species_id <- end_points[1]
  dest_species_id <- end_points[2]

  target_species_paths <- cpr$network$compartmentalize_species_pairs(
    consensus_model,
    source_species_id,
    dest_species_id
    )

  shortest_paths_list <- try(cpr$network$find_all_shortest_reaction_paths(
    consensus_graph,
    consensus_model,
    target_species_paths,
    weight_var = "weights"
    ),silent = TRUE)

  plot_shortest_path_network(
    shortest_paths_list,
    consensus_graph,
    max_labeled_species = 10L
    )
}
}
