% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viz_utils.R
\name{prepare_rendering}
\alias{prepare_rendering}
\title{Prepare Rendering}
\usage{
prepare_rendering(network, reaction_sources, network_layout)
}
\arguments{
\item{network}{a graphical network}

\item{reaction_sources}{an optional mapping from `r_id` to `pathway_id` and `name`.}

\item{network_layout}{method to used for creating a network layout (e.g., `fr`, `kk`, `drl`)}
}
\value{
a list containing
\describe{
    \item{neighborhood_grob}{A ggplot `grob` with no aethetics or geoms added}
    \item{vertices_df}{A table of vertices with upstream metadata plus coordinates and updated labels}
    \item{pathway_coords}{The bounding box of reaction's assigned to each pathway}
}
}
\description{
Prepare for the network plot by laying out the pathway in a way that loosly
groups related vertices and by suppressed overplotted labels
}
\keyword{internal}
