% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config_python.R
\name{setup_python_env}
\alias{setup_python_env}
\title{Setup Python Environment}
\usage{
setup_python_env(napistu_config, verbose = TRUE)
}
\arguments{
\item{napistu_config}{Napistu configuration object}

\item{verbose}{provide extra logging}
}
\value{
A `python_list` containing Python modules and environment info
}
\description{
Setup Python Environment
}
