% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_select_entity.R
\name{shiny_selectEntity_test}
\alias{shiny_selectEntity_test}
\title{Shiny Select Entity Test}
\usage{
shiny_selectEntity_test(species_names, species_identifiers)
}
\arguments{
\item{species_names}{A tibble containing names of all reaction species.}

\item{species_identifiers}{A tibble containing systematic identifiers for all reaction species nested by ontology.}
}
\value{
a shiny app
}
\description{
Demo for selecting an entity (species or reaction). Currently just species.
}
\examples{

if (interactive()) {
  interactive_initialization_wrapper()

  shiny_selectEntity_test(species_names, species_identifiers)
}
}
