% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_shortest_paths.R
\name{shiny_shortestPaths_test}
\alias{shiny_shortestPaths_test}
\title{Shiny Shortest Paths Test}
\usage{
shiny_shortestPaths_test(
  species_names,
  species_identifiers,
  consensus_model,
  consensus_graph,
  cpr
)
}
\arguments{
\item{species_names}{A tibble containing names of all reaction species.}

\item{species_identifiers}{A tibble containing systematic identifiers for all reaction species nested by ontology.}

\item{consensus_model}{A python pathway object with class cpr.sbml.SBML_dfs}

\item{consensus_graph}{A python igraph network of \code{consensus_model}}

\item{cpr}{reticulate connection to the cpr python package}
}
\value{
a shiny app
}
\description{
Demo for finding the shortest paths between pairs of entities.
}
\examples{

if (interactive()) {
  interactive_initialization_wrapper()
  cpr <- reticulate::import("cpr")

  shiny_shortestPaths_test(
    species_names,
    species_identifiers,
    consensus_model,
    consensus_graph,
    cpr = cpr
    )
}
}
