% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_shortest_paths.R
\name{shortestPathsServer}
\alias{shortestPathsServer}
\title{Shortest Paths Server}
\usage{
shortestPathsServer(
  id,
  species_names,
  species_identifiers,
  consensus_model,
  consensus_graph,
  cpr
)
}
\arguments{
\item{id}{An ID string that corresponds with the ID used to call the module's
UI function.}

\item{species_names}{A tibble containing names of all reaction species.}

\item{species_identifiers}{A tibble containing systematic identifiers for all reaction species nested by ontology.}

\item{consensus_model}{A python pathway object with class cpr.sbml.SBML_dfs}

\item{consensus_graph}{A python igraph network of \code{consensus_model}}

\item{cpr}{reticulate connection to the cpr python package}
}
\value{
Nothing; used for side-effects
}
\description{
Server-side components for the shortest paths shiny module
}
