% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/open_targets.R
\name{summarize_diseases}
\alias{summarize_diseases}
\title{Summarize Diseases}
\usage{
summarize_diseases(
  consensus_model,
  neighborhood_summary_table,
  species_identifiers
)
}
\arguments{
\item{consensus_model}{network reconstruction}

\item{neighborhood_summary_table}{tibble produced by
\code{create_neighborhood_summary_table}}

\item{species_identifiers}{A tibble containing systematic identifiers for all reaction species nested by ontology.}
}
\value{
A list containing targets and indications
}
\description{
Using a molecular neighborhood, target and indication summaries for all
  genes from Open Targets.
}
\examples{

if (interactive()) {
  interactive_initialization_wrapper()

  species_id <- rownames(consensus_model$species)[
    consensus_model$species$s_name == "KRAS:GTP"
  ]
  neighborhood_summary_table <- create_neighborhood_summary_table(
    create_neighborhood_table(
      species_id,
      consensus_model,
      consensus_graph,
      cpr = cpr,
      max_steps = 7L
    )
  )
  summarize_diseases(
    consensus_model,
    neighborhood_summary_table,
    species_identifiers
    )
}
}
