% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/open_targets.R
\name{summarize_diseases}
\alias{summarize_diseases}
\title{Summarize Diseases}
\usage{
summarize_diseases(napistu_list, neighborhood_summary_table)
}
\arguments{
\item{napistu_list}{A list containing loaded assets and bindings to
  Python modules.

\describe{
    \item{sbml_dfs}{
        SBML_dfs - the core pathway representation of the Napistu Python library
        }
    \item{napistu_graph}{
        Network graph - a Python igraph subclass with Napistu-specific
        attributes and methods
        }
    \item{species_identifiers}{Species identifier mappings}
    \item{precomputed_distances}{optional, distances between species nodes}
    \item{species_names}{
         A tibble containing the names of all genes, proteins, molecules, etc
         }
    \item{identifiers_nest}{
         A tibble with one row per ontology and a nested tibble containing
         all the identifiers and their corresponding molecular species
         }
    \item{python_modules}{A named list of Python modules: `napistu`}
    \item{python_environment}{See \link{validate_python_environment}}
    \item{napistu_config}{
        A `napistu_config` object dictating how the `napistu_list` was initialized
    }
    \item{loaded_at}{A date-time object indicating when `napistu_list` was initialized}
  }}

\item{neighborhood_summary_table}{tibble produced by
\code{create_neighborhood_summary_table}}
}
\value{
A list containing targets and indications
}
\description{
Using a molecular neighborhood, target and indication summaries for all
  genes from Open Targets.
}
\examples{
suppressPackageStartupMessages(library(dplyr))
setup_napistu_list(create_napistu_config())
species_id <- random_species(napistu_list)

neighborhood_summary_table <- create_neighborhood_table(
    napistu_list,
    species_id,
    max_steps = 7L
) \%>\%
    create_neighborhood_summary_table()

# TO DO - uncomment once https://github.com/napistu/napistu/issues/14 is addressed      
# summarize_diseases(
#        napistu_list,
#        neighborhood_summary_table
#    )
}
