% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/open_targets.R
\name{summarize_indication}
\alias{summarize_indication}
\title{Summarize Indications}
\usage{
summarize_indication(
  disease_id,
  sbml_dfs,
  neighborhood_summary_table,
  species_identifiers
)
}
\arguments{
\item{disease_id}{A disease to search from the EFO ontology}

\item{sbml_dfs}{network reconstruction}

\item{neighborhood_summary_table}{tibble produced by
\code{create_neighborhood_summary_table}}

\item{species_identifiers}{A tibble containing systematic identifiers for all reaction species nested by ontology.}
}
\value{
A tibble of species names and disease scores
}
\description{
Using a molecular neighborhood, find all Open Target scores for a specified disease.
}
\examples{

if (interactive()) {
  interactive_initialization_wrapper()

  species_id <- rownames(sbml_dfs$species)[
    sbml_dfs$species$s_name == "pyruvate kinase tetramer"
  ]
  neighborhood_summary_table <- create_neighborhood_summary_table(
    create_neighborhood_table(
      species_id,
      sbml_dfs,
      napistu_graph,
      napistu = napistu,
      max_steps = 7L
    )
  )
  disease_id <- "EFO_0000400" # diabetes
  summarize_indication(disease_id, sbml_dfs, neighborhood_summary_table, species_identifiers)
}
}
