% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shortest_paths.R
\name{summarize_shortest_paths}
\alias{summarize_shortest_paths}
\title{Summarize Shortest Paths}
\usage{
summarize_shortest_paths(
  source_species_id,
  dest_species_id,
  sbml_dfs,
  napistu_graph,
  napistu
)
}
\arguments{
\item{source_species_id}{species id to start with}

\item{dest_species_id}{species id to end at}

\item{sbml_dfs}{network reconstruction}

\item{napistu_graph}{igraph model of network}

\item{napistu}{binding to napistu python library}
}
\description{
Summarize Shortest Paths
}
\examples{

if (interactive()) {
  interactive_initialization_wrapper()

  # end_points <- sample(rownames(species_names), 2)
  # source_species_id <- end_points[1]
  # dest_species_id <- end_points[2]

  # source_species_id <- "S00002615"
  # dest_species_id <- "S00004885"

  source_species_id <- "S00002615"
  dest_species_id <- "S00000374"

  summarize_shortest_paths(
    source_species_id,
    dest_species_id,
    sbml_dfs,
    napistu_graph,
    napistu = napistu
  )
}
}
