% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config_utils.R
\name{validate_and_import_modules}
\alias{validate_and_import_modules}
\title{Validate and Import Python Modules}
\usage{
validate_and_import_modules(
  required_modules = NAPISTU_CONSTANTS$REQUIRED_PYTHON_MODULES
)
}
\arguments{
\item{required_modules}{a character vector of modules to import}
}
\value{
List of imported Python modules
}
\description{
Validate and Import Python Modules
}
