% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validators.R
\name{validate_assets_config}
\alias{validate_assets_config}
\title{Validate Assets Config}
\usage{
validate_assets_config(assets_config)
}
\arguments{
\item{assets_config}{List containing asset configuration settings, typically
extracted from the main napistu_config object}
}
\value{
Invisible TRUE if valid, throws error if invalid
}
\description{
Validate Assets Config
}
\keyword{internal}
