% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validators.R
\name{validate_napistu_config}
\alias{validate_napistu_config}
\title{Validate Napistu Configuration Object}
\usage{
validate_napistu_config(napistu_config)
}
\arguments{
\item{napistu_config}{A napistu configuration object of class 'napistu_config'
containing all necessary configuration settings for the loading required
assets and configuring and/or using an exiting appropriate Python environment}
}
\value{
Invisible TRUE if valid, throws error if invalid
}
\description{
Validate Napistu Configuration Object
}
\keyword{internal}
