% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validators.R
\name{validate_python_config}
\alias{validate_python_config}
\title{Validate Python Config}
\usage{
validate_python_config(python_config)
}
\arguments{
\item{python_config}{List describing whether to use the system python,
virtual environment or conda and what environment to use if using an
environment manager. Typically extracted from the main napistu_config
object.}
}
\value{
Invisible TRUE if valid, throws error if invalid
}
\description{
Validate Python Config
}
\keyword{internal}
